Refactor: 폴더명 변경 모달 공통모달로 교체
